requirements.txt
pyaudio
deepgram-sdk==3.2.7
google-generativeai
python-dotenv
keyboard
